Move Blaze promotions to DiscoverServerHandler
